add insertToTextDocument to editor
